Move more codes to util
